requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
psycopg2==2.8.6
Pygments==2.7.3
python-decouple==3.4
pytz==2020.5
sqlparse==0.4.1
